Reorganizes source code and refactors functions
To make the package more general and user friendly, we did away with the
Kafka Client struct and created functions to return either a consumer,
async producer or sync producer. It should now be much more simple for
the user; they only have to specify topic and consumer group or make a
simple function call to make a consumer or producer, respectively.

The source code was in a lower level within the directory (in the gb
fashion) which was messing up the imports. Now it is at the top level.