Removes AppConfig and dependency for handling env
This is the first of multiple commits to overhaul the package. The idea
is to let the user set their own configs instead of using a default.
This will be easier to use and be useful to more people.

In this first step, the AppConfig, which was simply holding the env
variables, is done away with and now the env vars are retrieved as needed
using the standard OS package instead of a third party package.

Basically, these are the barebones of what this package does. Next step
will be to build up the functionality and add in functions to create
consumers and producers.