Create a cluster consumer and async producer
This is the second step in the rewrite process. This serves the purposes
of the internal app using this package. The user can create a config to
their own specifications and this package handles the TLS portion of the
config, getting the broker addresses and appending the prefix where
needed.

The NewClient func may be done away with. At this point there is not a
clear reason to keep it, but that may change.

Next step is to create sync producers and sarama consumers.