Creates sarama consumer and sync producer
Adds some functionality for creating new consumer and producer types.
The sarama consumer may not be too helpful in the end, and it is unclear
how to/if it is possible or needed to set the consumer topic. We also
added some documentation for the public functions.

Next step will be to add quick start functions for creating consumers
and producers without supplying a config.