Expose NewTLSConfig for building the raw tls.Config with verification
The NewTLSConfig is, in some ways, the meat of what the library does and
hiding it behind the sugar of configs and env parsing forces anyone
using the library down a certain path.

I'm looking at using Viper for configuration on a project and want to
leverage it for pulling the environment variables. This is now possible
without copy and pasting code with NewTLSConfig exposed.